Control and mapping of mux_pos_num to mux_band_num
To make the mapping between the mux_pos_num to mux_band_num for each mux chip to be clear we are now reading this data from a csv file, an example of which is found in mux_pos_num_to_mux_band_num.csv. Added significant error handling and checking for when this file is read-in. Removed naked try statements to do with this file and its data to simplify the code for future users. Additionally, we removed some excess data not needed for the working example. Added a __name__ == '__main__' for the primary file, makemap.py. Suppressed some prints statements that were used for debugging try statement in read_iv.py, the entrire try statement was also removed, simplifying the code's logic. Started a ne configurations directory to start separating what is static code and what will need to be dynamically updated for each ufm.