detmap_conifg_example.py, now using absolute paths
The example is tested and working. The path mappings of files have been changes to have 3 directories that contain data, sample_data has large files that will change for every mapping but are too large to host on github, they will be downloaded automatically for the working example. metadata has static files that are the same for every mapping. config_files are smaller files that will change per-map, but should be committed to git as an example of the data contents and formatting. By dynamically determining the absolute paths to specify these file locations, the detmap code will continue to work when it is moved into another file structure.