New data structure, see channel_assignment.py
The data structures are being transitioned from Pandas data frames to a custom data structure written in pure python. This is done to move all the data mapping operations to a simplified central source. Pandas is a perfectly reasonable system for data handling, however it is not being used in a best practices manner, data frames lack a primary key (probably should be smurfband-channel_num tuples), there are a number of 0(n^2) process that can be implied to simplified to 0(n) mappings, integers are being cast as floats, and the column naming systems are inconsistent across several files of source code. While it is technically possible to do all these fixes within a pandas framework, it is the author's opinion that it will be easier to rewrite first in python and then export to pandas for Input/ouput, an additional pandas refactoring could also occur if users find that more readable compared to python.