Data Casting and Validation
An intermediate commit, while working on casting the Pandas' data frames into custom data class that do automatic data validations, correct value casting, and are faster to search and map. This commit is mostly needed so that I can save some progress and see what is changing and what is static in my editor as I continue to write.