Design Data Casting
The design data is imported from a pickle file and read directly into
Pandas. It has column names that with incorrect units, and misleading
column names (band, frequency-index) referring to mux chip
designations, as they are ambiguous when considered alongside
similarly named data for the SMuRF systems.

This design file is now automatically transformed into a
human-readable csv (Pandas-friendly). This csv file is the primary
input that code searches for, only using the pickle file when the
csv is not found. Changes to this file will propagate to data
processing steps in the rest of the code.

This file does not include smurf_band or channel columns, those
are assigned dynamically on read-in based on the frequency ordering
and smurf band definitions in vna_to_smurf.py.

This also represents a major timing and resource improvement. A
process used to take 160 second with an O(n^2) matching is now
takes less than 1 second to do a simple mapping.

The design data is held by the same class that does all the other
SMuRF tune data processing, so it comes with all the existing
validations, csv read/write, casting to Pandas data-frames,
interation support and everything else.