Preparation for Layout Data
To make ready for the Layout Topology, we now structure the data by
is_north (bool or None), then smurf_band (int), then channel_index
(int).

In this commit, I learned that smurf_channel was not what I assumed it
was. I actually wanted channel_index, basically a naive counter of
resonators in order within a smurf_band.

For SMuRF generated tune files, some resonators or (resonator-like
features) are flagged to not be used for readout, these are denoted
with smurf_channel == -1. These are currently, these are not being
matched to design frequencies. That choice will need to be revisited
and debated with a larger audience.