Layout data now available.
layout_data.py includes the tools needed for reading and parsing the
UXM layout metadata. Primary key of mux_layout_position, secondary
key of bond_pad. This uses the data format and exising special
handling, for example freq_obs_ghz.

Naming conventions were updated to be inline with the design
metadata and smurf tunefiles, with extra descriptors added to some
existing data columns to aid the user when the was ambiguity in names.

makemap.py was also updated and modularized to separate out tasks for
smurf, vna, design, and layout data. It is now much clearer what data
is needed for smurf-only or vna-only design and layout data mapping.

detector_map.py is now fully deprecated, and will be deleted in a
subsequent commit. channel_assignment.py is the effective
replacement.

Still, there is not a fully working example. What remains stage is to
update the plotting functions, this is the next target for upgrade.