Human-readable design file for collaboration
Only the pickle file version of this was committed in
the past. However, for ease of coloration and sharing,
I am now committing the csv file that was generated
as an intermediate step.

The pickle file is ignored when this file is available.
The pickle file will still be considered the primary
data source.