Changes requested by jlashner, from pull request
This does not address all the changes requested by Jack Lashner, some
changes are still in progress. The requested documentation will take
a few more days.

Changes this Commit
1) Metadata files all moved to detmap/meta/. These are files that
should be versioned, will not change often, but could change.

2) Jack asked about 'minimal example' of how to run detmap. The example
iPythonNotebook was moved to detmap/example/makemap.ipynb. Likely this
is what a new user would want. However, I like to run and debug based
on a 'compete test', what was under if __name__ == '__main__': in
detmap/makemap.py has been packaged into the definition psat_map
(use: from sodetlib.detmap.makemap import psat_map). A minimalist
version to run psat_map() is found in file: detmap/example/tune_map.py.
This file imports a single config variable and uses it to set the
kwargs of psat_map. This what can become an argparse script when
appropriate.

3) As requested, there is now detmap/examples/ dir.

4) I am not taking out the inline comments, even if advanced users find
them distracting. I have a lot of users on other projects, this is a
learned technique to reduce the number of emails I receive. A lot of
people never learn modular structure and are very confused data not in
square arrays.

5) All the design data (in detmap/meta/) has been converted to CSV to
be human-readable. Also in detmap/meta/, there is the primary data
structures I received at the start of the project, like pickle files.
The ability to convert from primary data to CSV is automatic, if a csv
file is not available the code looks for the pickle file and converts
to CSV and then reads the CSV.

6) The python file detmap_conifg_example.py is renamed to
detmap_config.py. Data that is changed often (tunfiles) or is user
based (psat_temperture, psat_show_plot) data was moved into a YAML file,
see detmap/example/example.yaml. The YAML file read-in and specified in
detmap_config.py. File basenames are in the YAML File but paths are
still handled in detmap_config.py using os.path.join(). I write all
my code to run cross-platform. detmap_config.py can be upgrade to
take paths from the YMAL file upon request.

7) As requested, I deleted the Dockerfile (and .dockerignore) from the
repo. I think it has a use-case, but it is so easy to write that we can
always add it back.

8) This is a bit of a sore spot, read_iv.py, is going to be deprecated
soon. It was my priority to read fresh data from Simons1 and test on
other data sets before my priorities were shifted from development to
export and user requests. As requested, it was moved to
detmap/example/read_iv.py.

9) timer_wrap.py, was used for development. It is deleted from the
repo.