Docstrings for makemap.py, bear-typing, simplification.
As requested in Jack Lashner's comments on the pull request for
the `caleb` branch, docstrings have been added to all the functions
in makemap.py. Numpy-style docstrings.

As a part of the organization, it was natural to implement bear-typing
to further constrain the functions to conform with what is written in
the docstrings.

Thing resulted in some consolidation and reorganization of the examples.
The examples are tested and up-to-date with the doc-strings in
makemap.py

Substantial work remains for added docstrings, namely in the primary
data class OperateTuneData within channel_assignment.py.