setup.py has been modified to include metadata.
The metadata CSV files in sodetlib/detmap/meta/ are needed for
as the default input for many of the functions in makemap.py.

In the future this metadata will be automatically chosen from the
data in this folder based on the name of the detector array for which
the mapping is being calculated. So this data will be relied on more
in the future, and will be used to create a simpler user experience.

Over-rides for custom metadata inputs will remain  available for the
most ambitious and experimental users.