More examples added based on user experiences.
I will continue to add demonstrations to sodetlib/detmap/examples
as my user base expands, and I get feedback on how users wish to
interact with the detmap module.

I hope to reduce files in this directory once it is know how and where
people use this system. For now, I will maintain a rather large files
set of examples.