Cleaned up code for Mapping Strategy 'map_by_freq'
The name space was a mess and the plotting program uninterpretable to
anyone but the designer. The per smurf-band part of the strategy was
moved to an independent definition to simply the namespace. The variable
names were renamed to be more independent and general, and no longer
specific to the larger strategies initial design.

Plots have been remade to have input and output measured data consistent
with a single step in the strategy, and not the overall data
transformation. Titles were added to the plots.

Verbose statements were added to inform which sections of the
strategy are triggered due to various data states.

Sections of the strategy are now skipped when they are not needed,
and would have no effect. This can be seen in which plots are shown and
what text is printed in the verbose statements.