New plot propagated to demos and docstrings.
Everything works again with access to new kwargs.

The get_config() definition in detmap_config.py now has a kwarg,
'output_data_dir', that allows the user to set a single output
directory for all plots and csv files. This is for users who do not
have write access to the directories with data, the default save
location.