Change in the handling of design metadata.
Mux-position on array is now assigned when design data is read-in.
This is to humiliate multiple assignments and repeated work when
multiple tunefiles use the same design data.

Metadata is now dynamically selected based on the array names
using get_config() in detmap_config.py. The old way of allowing the
user to select the metadata files was confusing and lead to the
creation of meaningless data that look like real results.
It is still easy to override this functionality, but this will
likely become less visible as the project continues to be developed
and packaged.

Changing how metadata is selected is controlled in the new file
meta_select.py.