Initial commit of a user interface restructuring.
I am restructuring the user interface with an additional class
layer that for running per-array batches of OperateTuneDate().

Why? Users are often operating on single arrays at a time but may
want to process a single tunefile or multiple tunefiles. The new
architecture allows this to happen with only a single loading of
the required design, layout, and assembly metadata.

The new per-array structure also allows us to add and
associate other per-array metadata, like array name, into the plots
and output files. This is a natural step for batch outputs across
tunefiles, and eventually across arrays without simply sending the
data to lower level functions to echo withun those more granular
architectures.