Updated and tested all the example demonstrations
Everything now works in the new framework.

I deleted the 'simple examples', I am only going to maintain the
demo_full examples and the examples that are custom to specific user
requests. This is still a growing body of demonstration code.
The simple examples were redundant, and a result of and early
user request that is no longer required.