Drop ipykernel and jupyterlab from requirements
JupyterLab and ipykernel are only used in the smurf-jupyter interactive
containers, and are not direct dependencies within sodetlib. 'jupyter' is also
already installed in [1], which pulls in both ipykernel and jupyterlab.

ipykernel pulls in ipython, which is actually used in start_pysmurf_ipython.py,
so that has been added here.

[1] - https://github.com/simonsobs/smurf_dockers/blob/main/pysmurf/requirements.txt