Drop extra dependencies (#483)
* Drop jedi from requirements.txt

jedi is a static analysis tool, helpful for developing, but not a direct
dependency of sodetlib [1].

[1] - https://pypi.org/project/jedi/

* Drop seaborn from requirements.txt

seaborn is not imported anywhere within sodetlib. If it's mainly used when
sodetlib is interactively used within a smurf-jupyter session then it is
already installed in the base docker image via [1].

[1] - https://github.com/simonsobs/smurf_dockers/blob/main/pysmurf/requirements.txt

* Drop pyfftw from requirements.txt

pyfftw is not imported anywhere in sodetlib. It seems like it maybe once was,
but was moved to sotodlib, i.e. in sotodlib.tod_ops.fft_ops.

* Drop ipykernel and jupyterlab from requirements

JupyterLab and ipykernel are only used in the smurf-jupyter interactive
containers, and are not direct dependencies within sodetlib. 'jupyter' is also
already installed in [1], which pulls in both ipykernel and jupyterlab.

ipykernel pulls in ipython, which is actually used in start_pysmurf_ipython.py,
so that has been added here.

[1] - https://github.com/simonsobs/smurf_dockers/blob/main/pysmurf/requirements.txt